Set up MariaDB database integration for the ERA server.
I have set up and tested MariaDB database integration with the ERA
Server. I implemented the stubbed DAOs, changed the command line options
to allow specifying a connection string, and added a new
`#uploadCourses` API to test the database functionality.

+++ server/src/main/java/era/server/data/model/Semester.java
@@ -30,7 +30,7 @@ public class Semester implements Comparable<Semester> {
         }
     }
 
-    private int uniqueId;
+    private long uniqueId;
     private Term term;
     private Year year;
 
@@ -41,7 +41,7 @@ public class Semester implements Comparable<Semester> {
         this.year = year;
     }
 
-    public Semester(int uniqueId, String term, @Nullable Integer year) {
+    public Semester(long uniqueId, String term, @Nullable Integer year) {
         Preconditions.checkArgument(uniqueId > 0, "A database id cannot be less than 1!");
         Preconditions.checkNotNull(term, "Term cannot be null!");
         this.uniqueId = uniqueId;
@@ -54,7 +54,7 @@ public class Semester implements Comparable<Semester> {
         return new Semester(term, year);
     }
 
-    public int getUniqueId() {
+    public long getUniqueId() {
         return uniqueId;
     }
